requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
async-timeout==5.0.1
asyncio-atexit==1.0.1
atomicwrites==1.4.0
attrs==25.3.0
backports.weakref==1.0.post1
black==25.1.0
branca==0.7.1
cachetools==5.3.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
constantly==15.1.0
cssselect==1.1.0
docstring-to-markdown==0.17
et-xmlfile==1.1.0
exceptiongroup==1.3.0
fastapi==0.116.1
folium==0.16.0
fonttools==4.25.0
frozenlist==1.7.0
google-api-core==2.19.1
google-api-python-client==2.135.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.2
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
jedi==0.19.2
Jinja2==3.1.6
jsonpointer==2.1
MarkupSafe==3.0.2
multidict==6.6.4
munkres==1.1.4
mypy_extensions==1.1.0
numpy==2.2.6
oauthlib==3.2.2
openpyxl==3.0.10
packaging==25.0
pandoc==2.4
parso==0.8.5
pathspec==0.12.1
patsy==0.5.3
pep8==1.7.1
pillow==11.3.0
platformdirs==4.4.0
playwright==1.55.0
pluggy==1.6.0
plumbum==1.9.0
ply==3.11
propcache==0.3.2
proto-plus==1.24.0
protobuf==5.27.2
py4j==0.10.9.7
pyairtable==3.0.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==2.10.2
pydantic_core==2.27.1
PyDispatcher==2.0.5
pyee==13.0.0
pyls-spyder==0.4.0
pyparsing==3.2.3
PyQt5-sip==12.11.0
pyspark==3.5.5
python-dotenv==1.1.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.13.1
python-multipart==0.0.20
queuelib==1.5.0
-e git+https://github.com/limgabriel/tiktok-valuereimagined.git@61d523cdec90ed688b67858d4ac21fdcd55df00e#egg=realitydefender&subdirectory=backend/aigc_api/realitydefender-sdk-python
regex==2025.7.34
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
TikTokApi==7.1.0
tiktoken==0.9.0
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.12.2
ujson==5.11.0
uritemplate==4.1.1
urllib3==2.5.0
uvicorn==0.35.0
validators==0.35.0
webencodings==0.5.1
xyzservices==2025.4.0
yarl==1.20.1
zipp==3.23.0
